Rename TranslatorUI to TranslateUI.
