+++ getting-started/background-processing/TranslateUI/Startup.cs
@@ -28,7 +28,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace TranslatorUI
+namespace TranslateUI
 {
     public class Startup
     {
