+++ getting-started/background-processing/TranslateUI/TranslateUI.csproj
