+++ getting-started/background-processing/TranslateUI/Views/Home/Index.cshtml
@@ -14,7 +14,7 @@
 // the License.
 }
 
-@model TranslatorUI.Models.HomeViewModel
+@model TranslateUI.Models.HomeViewModel
 @{
     ViewData["Title"] = "Home Page";
 }
