+++ getting-started/background-processing/TranslateUI/Views/_ViewImports.cshtml
@@ -1,4 +1,5 @@
-﻿// Copyright (c) 2019 Google LLC.
+﻿@{
+// Copyright (c) 2019 Google LLC.
 // 
 // Licensed under the Apache License, Version 2.0 (the "License"); you may not
 // use this file except in compliance with the License. You may obtain a copy of
@@ -11,15 +12,8 @@
 // WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 // License for the specific language governing permissions and limitations under
 // the License.
-
-using System;
-
-namespace TranslatorUI.Models
-{
-    public class ErrorViewModel
-    {
-        public string RequestId { get; set; }
-
-        public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
-    }
 }
+
+@using TranslateUI
+@using TranslateUI.Models
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
