+++ getting-started/background-processing/TranslateUI/wwwroot/js/site.js
