--- getting-started/background-processing/TranslatorUI/Models/HomeViewModel.cs
@@ -1,27 +1,0 @@
-﻿// Copyright (c) 2019 Google LLC.
-// 
-// Licensed under the Apache License, Version 2.0 (the "License"); you may not
-// use this file except in compliance with the License. You may obtain a copy of
-// the License at
-// 
-// http://www.apache.org/licenses/LICENSE-2.0
-// 
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
-// WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
-// License for the specific language governing permissions and limitations under
-// the License.
-
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-
-namespace TranslatorUI.Models
-{
-    public class HomeViewModel 
-    { 
-        public List<Translation> Translations { get; set; } = new List<Translation>();
-
-        [Required]
-        public string SourceText { get; set; }
-    }
-}
